split LuaRunner up so that Strings and Loader functions are separated
