+++ libs/server/Lua/LuaRunner.cs
@@ -144,610 +144,7 @@ namespace Garnet.server
                 curHead = origin + len;
             }
         }
-
-        /// <summary>
-        /// Just to DRY it up some, a holding type for all the constant strings we pre-load into
-        /// the Lua VM.
-        /// </summary>
-        private readonly struct ConstantStringRegistryIndexes
-        {
-            /// <see cref="CmdStrings.LUA_OK"/>
-            internal int Ok { get; }
-            /// <see cref="CmdStrings.LUA_ok"/>
-            internal int OkLower { get; }
-            /// <see cref="CmdStrings.LUA_err"/>
-            internal int Err { get; }
-            /// <see cref="CmdStrings.LUA_No_session_available"/>
-            internal int NoSessionAvailable { get; }
-            /// <see cref="CmdStrings.LUA_ERR_Please_specify_at_least_one_argument_for_this_redis_lib_call"/>
-            internal int PleaseSpecifyRedisCall { get; }
-            /// <see cref="CmdStrings.RESP_ERR_NOAUTH"/>
-            internal int ErrNoAuth { get; }
-            /// <see cref="CmdStrings.LUA_ERR_Unknown_Redis_command_called_from_script"/>
-            internal int ErrUnknown { get; }
-            /// <see cref="CmdStrings.LUA_ERR_Lua_redis_lib_command_arguments_must_be_strings_or_integers"/>
-            internal int ErrBadArg { get; }
-            /// <see cref="CmdStrings.LUA_ERR_wrong_number_of_arguments"/>
-            internal int ErrWrongNumberOfArgs { get; }
-            /// <see cref="CmdStrings.LUA_ERR_redis_log_requires_two_arguments_or_more"/>
-            internal int ErrRedisLogRequired { get; }
-            /// <see cref="CmdStrings.LUA_ERR_First_argument_must_be_a_number_log_level"/>
-            internal int ErrFirstArgMustBeNumber { get; }
-            /// <see cref="CmdStrings.LUA_ERR_Invalid_debug_level"/>
-            internal int ErrInvalidDebugLevel { get; }
-            /// <see cref="CmdStrings.LUA_ERR_Invalid_command_passed_to_redis_acl_check_cmd"/>
-            internal int ErrInvalidCommand { get; }
-            /// <see cref="CmdStrings.LUA_ERR_redis_setresp_requires_one_argument"/>
-            internal int ErrRedisSetRespArg { get; }
-            /// <see cref="CmdStrings.LUA_ERR_RESP_version_must_be_2_or_3"/>
-            internal int ErrRespVersion { get; }
-            /// <see cref="CmdStrings.LUA_ERR_redis_log_disabled"/>
-            internal int ErrLoggingDisabled { get; }
-            /// <see cref="CmdStrings.LUA_double"/>
-            internal int Double { get; }
-            /// <see cref="CmdStrings.LUA_map"/>
-            internal int Map { get; }
-            /// <see cref="CmdStrings.Lua_set"/>
-            internal int Set { get; }
-            /// <see cref="CmdStrings.LUA_big_number"/>
-            internal int BigNumber { get; }
-            /// <see cref="CmdStrings.LUA_format"/>
-            internal int Format { get; }
-            /// <see cref="CmdStrings.LUA_string"/>
-            internal int String { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_atan2"/>
-            internal int BadArgATan2 { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_cosh"/>
-            internal int BadArgCosh { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_frexp"/>
-            internal int BadArgFrexp { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_ldexp"/>
-            internal int BadArgLdexp { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_log10"/>
-            internal int BadArgLog10 { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_pow"/>
-            internal int BadArgPow { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_sinh"/>
-            internal int BadArgSinh { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_tanh"/>
-            internal int BadArgTanh { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_maxn"/>
-            internal int BadArgMaxn { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_loadstring"/>
-            internal int BadArgLoadString { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_loadstring_null_byte"/>
-            internal int BadArgLoadStringNullByte { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_tobit"/>
-            internal int BadArgToBit { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_tohex"/>
-            internal int BadArgToHex { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_bswap"/>
-            internal int BadArgBSwap { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_bnot"/>
-            internal int BadArgBNot { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_encode"/>
-            internal int BadArgEncode { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_decode"/>
-            internal int BadArgDecode { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_pack"/>
-            internal int BadArgPack { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_unpack"/>
-            internal int BadArgUnpack { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_bor"/>
-            internal int BadArgBOr { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_band"/>
-            internal int BadArgBAnd { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_bxor" />
-            internal int BadArgBXor { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_lshift"/>
-            internal int BadArgLShift { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_rshift"/>
-            internal int BadArgRShift { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_arshift"/>
-            internal int BadArgARShift { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_rol"/>
-            internal int BadArgRol { get; }
-            /// <see cref="CmdStrings.LUA_bad_arg_ror"/>
-            internal int BadArgRor { get; }
-            /// <see cref="CmdStrings.LUA_unexpected_json_value_kind"/>
-            internal int UnexpectedJsonValueKind { get; }
-            /// <see cref="CmdStrings.LUA_cannot_serialise_to_json"/>
-            internal int CannotSerialiseToJson { get; }
-            /// <see cref="CmdStrings.LUA_unexpected_error"/>
-            internal int UnexpectedError { get; }
-            /// <see cref="CmdStrings.LUA_cannot_serialise_excessive_nesting"/>
-            internal int CannotSerializeNesting { get; }
-            /// <see cref="CmdStrings.LUA_unable_to_format_number"/>
-            internal int UnableToFormatNumber { get; }
-            /// <see cref="CmdStrings.LUA_found_too_many_nested"/>
-            internal int FoundTooManyNested { get; }
-            /// <see cref="CmdStrings.LUA_expected_value_but_found_invalid"/>
-            internal int ExpectedValueButFound { get; }
-            /// <see cref="CmdStrings.LUA_missing_bytes_in_input"/>
-            internal int MissingBytesInInput { get; }
-            /// <see cref="CmdStrings.LUA_unexpected_msgpack_sigil"/>
-            internal int UnexpectedMsgPackSigil { get; }
-            /// <see cref="CmdStrings.LUA_msgpack_string_too_long"/>
-            internal int MsgPackStringTooLong { get; }
-            /// <see cref="CmdStrings.LUA_msgpack_array_too_long"/>
-            internal int MsgPackArrayTooLong { get; }
-            /// <see cref="CmdStrings.LUA_msgpack_map_too_long"/>
-            internal int MsgPackMapTooLong { get; }
-            /// <see cref="CmdStrings.LUA_insufficient_lua_stack_space"/>
-            internal int InsufficientLuaStackSpace { get; }
-            /// <see cref="CmdStrings.LUA_parameter_reset_failed_memory"/>
-            internal int ParameterResetFailedMemory { get; }
-            /// <see cref="CmdStrings.LUA_parameter_reset_failed_syntax"/>
-            internal int ParameterResetFailedSyntax { get; }
-            /// <see cref="CmdStrings.LUA_parameter_reset_failed_runtime"/>
-            internal int ParameterResetFailedRuntime { get; }
-            /// <see cref="CmdStrings.LUA_parameter_reset_failed_other"/>
-            internal int ParameterResetFailedOther { get; }
-            /// <see cref="CmdStrings.LUA_out_of_memory"/>
-            internal int OutOfMemory { get; }
-            /// <see cref="CmdStrings.LUA_load_string_error"/>
-            internal int LoadStringError { get; }
-            /// <see cref="CmdStrings.LUA_AND"/>
-            internal int AND { get; }
-            /// <see cref="CmdStrings.LUA_OR"/>
-            internal int OR { get; }
-            /// <see cref="CmdStrings.LUA_XOR"/>
-            internal int XOR { get; }
-            /// <see cref="CmdStrings.LUA_NOT"/>
-            internal int NOT { get; }
-            /// <see cref="CmdStrings.LUA_KEYS"/>
-            internal int KEYS { get; }
-            /// <see cref="CmdStrings.LUA_ARGV"/>
-            internal int ARGV { get; }
-
-            internal ConstantStringRegistryIndexes(ref LuaStateWrapper state)
-            {
-                // Commonly used strings, register them once so we don't have to copy them over each time we need them
-                //
-                // As an additional benefit, we don't have to worry about reserving memory for them either during normal operation
-                //
-                // This makes OOM probes unnecessary in many cases
-                Ok = ConstantStringToRegistry(ref state, CmdStrings.LUA_OK);
-                OkLower = ConstantStringToRegistry(ref state, CmdStrings.LUA_ok);
-                Err = ConstantStringToRegistry(ref state, CmdStrings.LUA_err);
-                NoSessionAvailable = ConstantStringToRegistry(ref state, CmdStrings.LUA_No_session_available);
-                PleaseSpecifyRedisCall = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_Please_specify_at_least_one_argument_for_this_redis_lib_call);
-                ErrNoAuth = ConstantStringToRegistry(ref state, CmdStrings.RESP_ERR_NOAUTH);
-                ErrUnknown = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_Unknown_Redis_command_called_from_script);
-                ErrBadArg = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_Lua_redis_lib_command_arguments_must_be_strings_or_integers);
-                ErrWrongNumberOfArgs = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_wrong_number_of_arguments);
-                ErrRedisLogRequired = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_redis_log_requires_two_arguments_or_more);
-                ErrFirstArgMustBeNumber = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_First_argument_must_be_a_number_log_level);
-                ErrInvalidDebugLevel = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_Invalid_debug_level);
-                ErrInvalidCommand = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_Invalid_command_passed_to_redis_acl_check_cmd);
-                ErrRedisSetRespArg = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_redis_setresp_requires_one_argument);
-                ErrRespVersion = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_RESP_version_must_be_2_or_3);
-                ErrLoggingDisabled = ConstantStringToRegistry(ref state, CmdStrings.LUA_ERR_redis_log_disabled);
-                Double = ConstantStringToRegistry(ref state, CmdStrings.LUA_double);
-                Map = ConstantStringToRegistry(ref state, CmdStrings.LUA_map);
-                Set = ConstantStringToRegistry(ref state, CmdStrings.Lua_set);
-                BigNumber = ConstantStringToRegistry(ref state, CmdStrings.LUA_big_number);
-                Format = ConstantStringToRegistry(ref state, CmdStrings.LUA_format);
-                String = ConstantStringToRegistry(ref state, CmdStrings.LUA_string);
-                BadArgATan2 = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_atan2);
-                BadArgCosh = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_cosh);
-                BadArgFrexp = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_frexp);
-                BadArgLdexp = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_ldexp);
-                BadArgLog10 = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_log10);
-                BadArgPow = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_pow);
-                BadArgSinh = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_sinh);
-                BadArgTanh = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_tanh);
-                BadArgMaxn = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_maxn);
-                BadArgLoadString = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_loadstring);
-                BadArgLoadStringNullByte = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_loadstring_null_byte);
-                BadArgToBit = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_tobit);
-                BadArgToHex = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_tohex);
-                BadArgBSwap = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_bswap);
-                BadArgBNot = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_bnot);
-                BadArgEncode = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_encode);
-                BadArgDecode = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_decode);
-                BadArgPack = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_pack);
-                BadArgUnpack = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_unpack);
-                BadArgBOr = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_bor);
-                BadArgBAnd = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_band);
-                BadArgBXor = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_bxor);
-                BadArgLShift = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_lshift);
-                BadArgARShift = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_arshift);
-                BadArgRShift = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_rshift);
-                BadArgRol = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_rol);
-                BadArgRor = ConstantStringToRegistry(ref state, CmdStrings.LUA_bad_arg_ror);
-                UnexpectedJsonValueKind = ConstantStringToRegistry(ref state, CmdStrings.LUA_unexpected_json_value_kind);
-                CannotSerialiseToJson = ConstantStringToRegistry(ref state, CmdStrings.LUA_cannot_serialise_to_json);
-                UnexpectedError = ConstantStringToRegistry(ref state, CmdStrings.LUA_unexpected_error);
-                CannotSerializeNesting = ConstantStringToRegistry(ref state, CmdStrings.LUA_cannot_serialise_excessive_nesting);
-                UnableToFormatNumber = ConstantStringToRegistry(ref state, CmdStrings.LUA_unable_to_format_number);
-                FoundTooManyNested = ConstantStringToRegistry(ref state, CmdStrings.LUA_found_too_many_nested);
-                ExpectedValueButFound = ConstantStringToRegistry(ref state, CmdStrings.LUA_expected_value_but_found_invalid);
-                MissingBytesInInput = ConstantStringToRegistry(ref state, CmdStrings.LUA_missing_bytes_in_input);
-                UnexpectedMsgPackSigil = ConstantStringToRegistry(ref state, CmdStrings.LUA_unexpected_msgpack_sigil);
-                MsgPackStringTooLong = ConstantStringToRegistry(ref state, CmdStrings.LUA_msgpack_string_too_long);
-                MsgPackArrayTooLong = ConstantStringToRegistry(ref state, CmdStrings.LUA_msgpack_array_too_long);
-                MsgPackMapTooLong = ConstantStringToRegistry(ref state, CmdStrings.LUA_msgpack_map_too_long);
-                InsufficientLuaStackSpace = ConstantStringToRegistry(ref state, CmdStrings.LUA_insufficient_lua_stack_space);
-                ParameterResetFailedMemory = ConstantStringToRegistry(ref state, CmdStrings.LUA_parameter_reset_failed_memory);
-                ParameterResetFailedSyntax = ConstantStringToRegistry(ref state, CmdStrings.LUA_parameter_reset_failed_syntax);
-                ParameterResetFailedRuntime = ConstantStringToRegistry(ref state, CmdStrings.LUA_parameter_reset_failed_runtime);
-                ParameterResetFailedOther = ConstantStringToRegistry(ref state, CmdStrings.LUA_parameter_reset_failed_other);
-                OutOfMemory = ConstantStringToRegistry(ref state, CmdStrings.LUA_out_of_memory);
-                LoadStringError = ConstantStringToRegistry(ref state, CmdStrings.LUA_load_string_error);
-                AND = ConstantStringToRegistry(ref state, CmdStrings.LUA_AND);
-                OR = ConstantStringToRegistry(ref state, CmdStrings.LUA_OR);
-                XOR = ConstantStringToRegistry(ref state, CmdStrings.LUA_XOR);
-                NOT = ConstantStringToRegistry(ref state, CmdStrings.LUA_NOT);
-                KEYS = ConstantStringToRegistry(ref state, CmdStrings.LUA_KEYS);
-                ARGV = ConstantStringToRegistry(ref state, CmdStrings.LUA_ARGV);
-            }
-
-            /// <summary>
-            /// Some strings we use a bunch, and copying them to Lua each time is wasteful
-            ///
-            /// So instead we stash them in the Registry and load them by index
-            /// </summary>
-            private static int ConstantStringToRegistry(ref LuaStateWrapper state, ReadOnlySpan<byte> str)
-            {
-                if (!state.TryPushBuffer(str))
-                {
-                    throw new GarnetException("Insufficient space in Lua VM for constant string");
-                }
-
-                return state.UnsafeRef();
-            }
-        }
-
-        /// <summary>
-        /// Simple cache of allowed functions to loader block.
-        /// </summary>
-        private sealed record LoaderBlockCache(HashSet<string> AllowedFunctions, ReadOnlyMemory<byte> LoaderBlockBytes);
-
-        private const string LoaderBlock = @"
--- disable for sandboxing purposes
-import = function () end
-
--- unpack moved after Lua 5.1, this provides Redis compat
-local unpack = table.unpack
-
--- common functions for handling error replies from Garnet
-local chain_func = function(f1, f2)
-    return function(...)
-        return f1(f2, ...)
-    end
-end
-
-local error_wrapper_r0 = function(rawFunc, ...) 
-    local err = rawFunc(...)
-    if err then
-        error(err, 0)
-    end
-end
-
-local error_wrapper_r1 = function(rawFunc, ...) 
-    local r1, err = rawFunc(...)
-    if err then
-        error(err, 0)
-    end
-
-    return r1
-end
-
-local error_wrapper_r2 = function(rawFunc, ...) 
-    local r1, r2, err = rawFunc(...)
-    if err then
-        error(err, 0)
-    end
-
-    return r1, r2
-end
-
-local unpackTrampolineRef = garnet_unpack_trampoline
-local error_wrapper_rvar_check = function(err, ...)
-    if err then
-        error(err, 0)
-    end
-
-    return ...
-end
-
-local error_wrapper_rvar = function(rawFunc, ...)
-    -- variable numbers of returns require extra work
-    -- and requires that the error token is FIRST, not last
-    -- and a count of expected returns is around to handle
-    -- trailing nils
-    local rets = {rawFunc(...)}
-    local err = rets[1]
-    local count = rets[2]
-    if err then
-        error(err, 0)
-    end
-
-    return error_wrapper_rvar_check(unpackTrampolineRef(rets, count))
-end
-
--- cutdown os for sandboxing purposes
-local osClockRef = os.clock
-os = {
-    clock = osClockRef
-}
-
--- define cjson for (optional) inclusion into sandbox_env
-local cjson = {
-    encode = chain_func(error_wrapper_r1, garnet_cjson_encode);
-    decode = chain_func(error_wrapper_r1, garnet_cjson_decode);
-}
-
--- define bit for (optional) inclusion into sandbox_env
-local garnetBitopRef = chain_func(error_wrapper_r1, garnet_bitop)
-local bit = {
-    tobit = chain_func(error_wrapper_r1, garnet_bit_tobit);
-    tohex = chain_func(error_wrapper_r1, garnet_bit_tohex);
-    bnot = function(...) return garnetBitopRef(0, ...); end;
-    bor = function(...) return garnetBitopRef(1, ...); end;
-    band = function(...) return garnetBitopRef(2, ...); end;
-    bxor = function(...) return garnetBitopRef(3, ...); end;
-    lshift = function(...) return garnetBitopRef(4, ...); end;
-    rshift = function(...) return garnetBitopRef(5, ...); end;
-    arshift = function(...) return garnetBitopRef(6, ...); end;
-    rol = function(...) return garnetBitopRef(7, ...); end;
-    ror = function(...) return garnetBitopRef(8, ...); end;
-    bswap = chain_func(error_wrapper_r1, garnet_bit_bswap);
-}
-
--- define cmsgpack for (optional) inclusion into sandbox_env
-local cmsgpack = {
-    pack = chain_func(error_wrapper_r1, garnet_cmsgpack_pack);
-    unpack = chain_func(error_wrapper_rvar, garnet_cmsgpack_unpack);
-}
-
--- define struct for (optional) inclusion into sandbox_env
-local struct = {
-    pack = string.pack;
-    unpack = string.unpack;
-    size = string.packsize;
-}
-
--- define redis for (optional, but almost always) inclusion into sandbox_env
-local garnetCallRef = chain_func(error_wrapper_r1, garnet_call)
-local pCallRef = pcall
-local redis = {
-    status_reply = function(text)
-        return text
-    end,
-
-    error_reply = function(text)
-        return { err = 'ERR ' .. text }
-    end,
-
-    call = garnetCallRef,
-
-    pcall = function(...)
-        local success, errOrRes = pCallRef(garnetCallRef, ...)
-        if success then
-            return errOrRes
-        end
-
-        return { err = errOrRes }
-    end,
-
-    sha1hex = chain_func(error_wrapper_r1, garnet_sha1hex),
-
-    LOG_DEBUG = 0,
-    LOG_VERBOSE = 1,
-    LOG_NOTICE = 2,
-    LOG_WARNING = 3,
-
-    log = chain_func(error_wrapper_r0, garnet_log),
-
-    REPL_ALL = 3,
-    REPL_AOF = 1,
-    REPL_REPLICA = 2,
-    REPL_SLAVE = 2,
-    REPL_NONE = 0,
-
-    set_repl = function(...)
-        -- this is a giant footgun, straight up not implementing it
-        error('ERR redis.set_repl is not supported in Garnet', 0)
-    end,
-
-    replicate_commands = function(...)
-        return true
-    end,
-
-    breakpoint = function(...)
-        -- this is giant and weird, not implementing
-        error('ERR redis.breakpoint is not supported in Garnet', 0)
-    end,
-
-    debug = function(...)
-        -- this is giant and weird, not implementing
-        error('ERR redis.debug is not supported in Garnet', 0)
-    end,
-
-    acl_check_cmd = chain_func(error_wrapper_r1, garnet_acl_check_cmd),
-    setresp = chain_func(error_wrapper_r0, garnet_setresp),
-
-    REDIS_VERSION = garnet_REDIS_VERSION,
-    REDIS_VERSION_NUM = garnet_REDIS_VERSION_NUM
-}
-
--- added after Lua 5.1, removing to maintain Redis compat
-string.pack = nil
-string.unpack = nil
-string.packsize = nil
-math.maxinteger = nil
-math.type = nil
-math.mininteger = nil
-math.tointeger = nil
-math.ult = nil
-table.pack = nil
-table.unpack = nil
-table.move = nil
-
--- in Lua 5.1 but not 5.4, so implemented on the .NET side
-local loadstring = chain_func(error_wrapper_r2, garnet_loadstring)
-math.atan2 = chain_func(error_wrapper_r1, garnet_atan2)
-math.cosh = chain_func(error_wrapper_r1, garnet_cosh)
-math.frexp = chain_func(error_wrapper_r2, garnet_frexp)
-math.ldexp = chain_func(error_wrapper_r1, garnet_ldexp)
-math.log10 = chain_func(error_wrapper_r1, garnet_log10)
-math.pow = chain_func(error_wrapper_r1, garnet_pow)
-math.sinh = chain_func(error_wrapper_r1, garnet_sinh)
-math.tanh = chain_func(error_wrapper_r1, garnet_tanh)
-table.maxn = chain_func(error_wrapper_r1, garnet_maxn)
-
-local collectgarbageRef = collectgarbage
-local setMetatableRef = setmetatable
-local rawsetRef = rawset
-
--- prevent modification to metatables for readonly tables
--- Redis accomplishes this by patching Lua, we'd rather ship
--- vanilla Lua and do it in code
-local setmetatable = function(table, metatable)
-    if table and table.__readonly then
-        error('Attempt to modify a readonly table', 0)
-    end
-
-    return setMetatableRef(table, metatable)
-end
-
--- prevent bypassing metatables to update readonly tables
--- as above, Redis prevents this with a patch to Lua
-local rawset = function(table, key, value)
-    if table and table.__readonly then
-        error('Attempt to modify a readonly table', 0)
-    end
-
-    return rawsetRef(table, key, value)
-end
-
--- technically deprecated in 5.1, but available in Redis
--- this is only 'sort of' correct as 5.4 doesn't expose the same
--- gc primitives
-local gcinfo = function()
-    return collectgarbageRef('count'), 0
-end
-
--- global object used for the sandbox environment
---
--- replacements are performed before VM initialization
--- to allow configuring available functions
-sandbox_env = {
-    _VERSION = _VERSION;
-
-    KEYS = KEYS;
-    ARGV = ARGV;
-
-!!SANDBOX_ENV REPLACEMENT TARGET!!
-}
--- timeout error must be raised on Lua
-local debugRef = debug
-local force_timeout = function()
-    error('ERR Lua script exceeded configured timeout', 0)
-end
-function request_timeout()
-    debugRef.sethook(force_timeout, '', 1)
-end
--- no reference to outermost set of globals (_G) should survive sandboxing
-sandbox_env._G = sandbox_env
--- lock down a table, recursively doing the same to all table members
-local rawGetRef = rawget
-local readonly_metatable = {
-    __index = function(onTable, key)
-        return rawGetRef(onTable, key)
-    end,
-    __newindex = function(onTable, key, value)
-        error('Attempt to modify a readonly table', 0)
-    end
-}
-function recursively_readonly_table(table)
-    if table.__readonly then
-        return table
-    end
-
-    table.__readonly = true
-
-    for key, value in pairs(table) do
-        if type(value) == 'table' then
-            recursively_readonly_table(value)
-        end
-    end
-
-    setMetatableRef(table, readonly_metatable)
-end
--- do resets in the Lua side to minimize pinvokes
-function reset_keys_and_argv(fromKey, fromArgv)
-    local keyRef = sandbox_env.KEYS
-    local keyCount = #keyRef
-    for i = fromKey, keyCount do
-        table.remove(keyRef)
-    end
-
-    local argvRef = sandbox_env.ARGV
-    local argvCount = #argvRef
-    for i = fromArgv, argvCount do
-        table.remove(argvRef)
-    end
-end
--- force new 'global' environment to be readonly
-recursively_readonly_table(sandbox_env)
--- responsible for sandboxing user provided code
-function load_sandboxed(source)
-    local rawFunc, err = load(source, nil, nil, sandbox_env)
-
-    return err, rawFunc
-end
-";
-        private static readonly HashSet<string> DefaultAllowedFunctions = [
-            // Built ins
-            "assert",
-            "collectgarbage",
-            "coroutine",
-            "error",
-            "gcinfo",
-            // Intentionally not supporting getfenv, as it's too weird to backport to Lua 5.4
-            "getmetatable",
-            "ipairs",
-            "load",
-            "loadstring",
-            "math",
-            "next",
-            "pairs",
-            "pcall",
-            "rawequal",
-            "rawget",
-            // Note rawset is proxied to implement readonly tables
-            "rawset",
-            "select",
-            // Intentionally not supporting setfenv, as it's too weird to backport to Lua 5.4
-            // Note setmetatable is proxied to implement readonly tables
-            "setmetatable",
-            "string",
-            "table",
-            "tonumber",
-            "tostring",
-            "type",
-            // Note unpack is actually table.unpack, and defined in the loader block
-            "unpack",
-            "xpcall",
-
-            // Runtime libs
-            "bit",
-            "cjson",
-            "cmsgpack",
-            // Note os only contains clock due to definition in the loader block
-            "os",
-            // Note struct is actually implemented by Lua 5.4's string.pack/unpack/size
-            "struct",
-
-            // Interface force communicating back with Garnet
-            "redis",
-        ];
-
-        private static LoaderBlockCache CachedLoaderBlock;
-
+                
         private static (int Start, ulong[] ByteMask) NoScriptDetails = InitializeNoScriptDetails();
 
         // References into Registry on the Lua side
@@ -3211,77 +2608,6 @@ end
             }
 
             return (start, bitmap);
-        }
-
-        /// <summary>
-        /// Modifies <see cref="LoaderBlock"/> to account for <paramref name="allowedFunctions"/>, and converts to bytes.
-        /// 
-        /// Provided as an optimization, as often this can be memoized.
-        /// </summary>
-        private static ReadOnlyMemory<byte> PrepareLoaderBlockBytes(HashSet<string> allowedFunctions)
-        {
-            // If nothing is explicitly allowed, fallback to our defaults
-            if (allowedFunctions.Count == 0)
-            {
-                allowedFunctions = DefaultAllowedFunctions;
-            }
-
-            // Most of the time this list never changes, so reuse the work
-            var cache = CachedLoaderBlock;
-            if (cache != null && ReferenceEquals(cache.AllowedFunctions, allowedFunctions))
-            {
-                return cache.LoaderBlockBytes;
-            }
-
-            // Build the subset of a Lua table where we export all these functions
-            var wholeIncludes = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-            var replacement = new StringBuilder();
-            foreach (var wholeRef in allowedFunctions.Where(static x => !x.Contains('.')))
-            {
-                if (!DefaultAllowedFunctions.Contains(wholeRef))
-                {
-                    // Skip functions not intentionally exported
-                    continue;
-                }
-
-                _ = replacement.AppendLine($"    {wholeRef}={wholeRef};");
-                _ = wholeIncludes.Add(wholeRef);
-            }
-
-            // Partial includes (ie. os.clock) need special handling
-            var partialIncludes = allowedFunctions.Where(static x => x.Contains('.')).Select(static x => (Leading: x[..x.IndexOf('.')], Trailing: x[(x.IndexOf('.') + 1)..]));
-            foreach (var grouped in partialIncludes.GroupBy(static t => t.Leading, StringComparer.OrdinalIgnoreCase))
-            {
-                if (wholeIncludes.Contains(grouped.Key))
-                {
-                    // Including a subset of something included in whole doesn't affect things
-                    continue;
-                }
-
-                if (!DefaultAllowedFunctions.Contains(grouped.Key))
-                {
-                    // Skip functions not intentionally exported
-                    continue;
-                }
-
-                _ = replacement.AppendLine($"    {grouped.Key}={{");
-                foreach (var part in grouped.Select(static t => t.Trailing).Distinct().OrderBy(static t => t))
-                {
-                    _ = replacement.AppendLine($"        {part}={grouped.Key}.{part};");
-                }
-                _ = replacement.AppendLine("    };");
-            }
-
-            var decl = replacement.ToString();
-            var finalLoaderBlock = LoaderBlock.Replace("!!SANDBOX_ENV REPLACEMENT TARGET!!", decl);
-
-            // Save off for next caller
-            //
-            // Inherently race-y, but that's fine - worst case we do a little extra work
-            var newCache = new LoaderBlockCache(allowedFunctions, Encoding.UTF8.GetBytes(finalLoaderBlock));
-            CachedLoaderBlock = newCache;
-
-            return newCache.LoaderBlockBytes;
         }
     }
 }
